KAF-549: refactored and renamed integration tests.
